[conluz-216] Moved shared agreement classes from plant generic package to a sibling sharingagreement specific package
