+++ src/main/java/org/lucoenergia/conluz/domain/production/plant/sharingagreement/SharingAgreementNotFoundException.java
@@ -1,4 +1,4 @@
-package org.lucoenergia.conluz.domain.production.plant;
+package org.lucoenergia.conluz.domain.production.plant.sharingagreement;
 
 import java.util.UUID;
