+++ src/main/java/org/lucoenergia/conluz/domain/production/plant/sharingagreement/SharingAgreementStatus.java
@@ -1,4 +1,4 @@
-package org.lucoenergia.conluz.domain.production.plant;
+package org.lucoenergia.conluz.domain.production.plant.sharingagreement;
 
 public enum SharingAgreementStatus {
     DRAFT,
